moved paypal provider to App
